support static model export

+++ paddlenlp/ops/fast_transformer/src/fusion_decoding_op.h
@@ -79,6 +79,6 @@ std::vector<paddle::Tensor> DecodingCUDAForward(
     const int& num_layer,
     const int& bos_id,
     const int& eos_id,
-    const int64_t& max_len,
+    const int& max_len,
     const float& beam_search_diversity_rate,
     const float& alpha);
