+++ paddlenlp/ops/fast_transformer/src/fusion_mbart_decoding_op.h
@@ -81,7 +81,7 @@ std::vector<paddle::Tensor> MBartDecodingCUDAForward(
     const int& start_id_,
     const int& end_id_,
     const float& temperature,
-    const int64_t& max_seq_len_,
+    const int& max_seq_len_,
     const float& beam_search_diversity_rate_,
     const float& alpha,
     const bool& early_stopping,
