+++ paddlenlp/ops/fast_transformer/src/fusion_t5_decoding_op.h
@@ -80,7 +80,7 @@ std::vector<paddle::Tensor> T5DecodingCUDAForward(
     const int& num_layer,
     const int& bos_id,
     const int& eos_id,
-    const int64_t& max_len,
+    const int& max_len,
     const float& beam_search_diversity_rate,
     const float& alpha,
     const float& temperature,
